Atualizações no SystemOperation.java

+++ src/so/SystemOperation.java
@@ -3,35 +3,34 @@ package so;
 import soCpu.CpuManager;
 import soMemory.MemoryManager;
 import soMemory.Strategy;
-import so.Process;
-import so.SystemCallType;
+import soSchedule.Schedule;
 
 public class SystemOperation {
     private static MemoryManager mm;
     private static CpuManager cm;
+    private static Schedule sc;
 
-    private static void createProcess() {
+    private static void startsMmAndCpu() {
         if (mm == null) {
-            mm = new MemoryManager(Strategy.BEST_FIT);
+            mm = new MemoryManager(Strategy.FIRST_FIT);
         }
         if (cm == null) {
             cm = new CpuManager();
         }
     }
 
-    public static void SystemCall(SystemCallType type, Process p) {
-        switch (type) {
-            case DELETE:
-                mm.deleteProcess(p);
-                break;
-            case READ:
-                mm.readProcess(p);
-                break;
-            case WRITE:
-                mm.writeProcess(p);
-                break;
-            default:
-                break;
+    public static Process systemCall(SystemCallType type, Process p) {
+        if (type.equals((SystemCallType.WRITE))) {
+             mm.writeProcess(p);
+        } else if (type.equals((SystemCallType.READ))) {
+            mm.readProcess(p);
+        } else if (type.equals((SystemCallType.CREATE))) {
+            startsMmAndCpu();
+            Process process = cm.createProcess(p);
+            return process;
+        } else if (type.equals((SystemCallType.DELETE))) {
+            mm.deleteProcess(p);  
         }
+        return null;
     }
 }
